Hackathon2-phase3/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "openai-agents>=0.1.0",
    "mcp>=0.1.0",
]
